fix: type err
