+++ packages/features/bookings/lib/handleNewBooking.ts
@@ -251,10 +251,15 @@ async function handler(
     })
   );
 
+  const user = eventType.users.find((user) => user.id === eventType.userId);
+  const userSchedule = user?.schedules.find((schedule) => schedule.id === user?.defaultScheduleId);
+  const eventTimeZone = eventType.schedule?.timeZone ?? userSchedule?.timeZone;
+
   await validateBookingTimeIsNotOutOfBounds<typeof eventType>(
     reqBody.start,
     reqBody.timeZone,
     eventType,
+    eventTimeZone,
     loggerWithEventDetails
   );
